+++ tests/compas_timber/test_connections_solvers.py
@@ -54,11 +54,11 @@ def test_find_topology_i_beams(i_beams):
             for beam_b in [bb, flip_beam(bb)]:
                 cs = ConnectionSolver()
                 topo_results = cs.find_topology(beam_a, beam_b)
-                assert topo_results[0] == JointTopology.TOPO_I, "Expected I-beam joint topology"
-                assert topo_results[1] == beam_a, "Expected first beam as first plate in topology result"
-                assert topo_results[2] == beam_b, "Expected second beam as second plate in topology result"
-                assert TOL.is_close(topo_results[3], 0.0), "Expected distance == 0"
-                assert isinstance(topo_results[4], Point), "Expected location to be a Point instance"
+                assert topo_results.topology == JointTopology.TOPO_I, "Expected I-beam joint topology"
+                assert topo_results.beam_a == beam_a, "Expected first beam as first plate in topology result"
+                assert topo_results.beam_b == beam_b, "Expected second beam as second plate in topology result"
+                assert TOL.is_close(topo_results.distance, 0.0), "Expected distance == 0"
+                assert isinstance(topo_results.location, Point), "Expected location to be a Point instance"
 
 
 def test_find_topology_tolerance_no_max_dist(i_beams_with_tolerance):
@@ -67,8 +67,8 @@ def test_find_topology_tolerance_no_max_dist(i_beams_with_tolerance):
             for beam_b in [bb, flip_beam(bb)]:
                 cs = ConnectionSolver()
                 topo_results = cs.find_topology(beam_a, beam_b)
-                assert topo_results[0] == JointTopology.TOPO_UNKNOWN, "Expected unknown joint topology"
-                assert all([v is None for v in topo_results[1:]]), "Expected no valid topology results for I_Topo beams with tolerance"
+                assert topo_results.topology == JointTopology.TOPO_UNKNOWN, "Expected unknown joint topology"
+                assert all([v is None for v in [topo_results.distance, topo_results.location]]), "Expected no valid topology results for I_Topo beams with tolerance"
 
 
 def test_find_topology_i_beams_with_max_dist(i_beams_with_tolerance):
@@ -77,11 +77,11 @@ def test_find_topology_i_beams_with_max_dist(i_beams_with_tolerance):
             for beam_b in [bb, flip_beam(bb)]:
                 cs = ConnectionSolver()
                 topo_results = cs.find_topology(beam_a, beam_b, max_distance=0.02)
-                assert topo_results[0] == JointTopology.TOPO_I, "Expected I-beam joint topology"
-                assert topo_results[1] == beam_a, "Expected first beam as first plate in topology result"
-                assert topo_results[2] == beam_b, "Expected second beam as second plate in topology result"
-                assert TOL.is_close(topo_results[3], 0.01), "Expected distance == 0.001"
-                assert isinstance(topo_results[4], Point), "Expected location to be a Point instance"
+                assert topo_results.topology == JointTopology.TOPO_I, "Expected I-beam joint topology"
+                assert topo_results.beam_a == beam_a, "Expected first beam as first plate in topology result"
+                assert topo_results.beam_b == beam_b, "Expected second beam as second plate in topology result"
+                assert TOL.is_close(topo_results.distance, 0.01), "Expected distance == 0.001"
+                assert isinstance(topo_results.location, Point), "Expected location to be a Point instance"
 
 
 def test_find_topology_i_beams_overlap(overlap_i_beams):
@@ -90,8 +90,8 @@ def test_find_topology_i_beams_overlap(overlap_i_beams):
             for beam_b in [bb, flip_beam(bb)]:
                 cs = ConnectionSolver()
                 topo_results = cs.find_topology(beam_a, beam_b)
-                assert topo_results[0] == JointTopology.TOPO_UNKNOWN, "Expected unknown joint topology"
-                assert all([v is None for v in topo_results[1:]]), "Expected no valid topology results for I_Topo beams with tolerance"
+                assert topo_results.topology == JointTopology.TOPO_UNKNOWN, "Expected unknown joint topology"
+                assert all([v is None for v in [topo_results.distance, topo_results.location]]), "Expected no valid topology results for I_Topo beams with tolerance"
 
 
 def test_find_topology_i_beams_overlap_max_dist(overlap_i_beams):
@@ -100,8 +100,8 @@ def test_find_topology_i_beams_overlap_max_dist(overlap_i_beams):
             for beam_b in [bb, flip_beam(bb)]:
                 cs = ConnectionSolver()
                 topo_results = cs.find_topology(beam_a, beam_b, max_distance=0.02)
-                assert topo_results[0] == JointTopology.TOPO_UNKNOWN, "Expected unknown joint topology"
-                assert all([v is None for v in topo_results[1:]]), "Expected no valid topology results for I_Topo beams with tolerance"
+                assert topo_results.topology == JointTopology.TOPO_UNKNOWN, "Expected unknown joint topology"
+                assert all([v is None for v in [ topo_results.distance, topo_results.location]]), "Expected no valid topology results for I_Topo beams with tolerance"
 
 
 @pytest.fixture
@@ -132,11 +132,11 @@ def test_find_topology_l_beams(l_beams):
             for beam_b in [bb, flip_beam(bb)]:
                 cs = ConnectionSolver()
                 topo_results = cs.find_topology(beam_a, beam_b)
-                assert topo_results[0] == JointTopology.TOPO_L, "Expected I-beam joint topology"
-                assert topo_results[1] == beam_a, "Expected first beam as first plate in topology result"
-                assert topo_results[2] == beam_b, "Expected second beam as second plate in topology result"
-                assert TOL.is_close(topo_results[3], 0.0), "Expected distance == 0"
-                assert isinstance(topo_results[4], Point), "Expected location to be a Point instance"
+                assert topo_results.topology == JointTopology.TOPO_L, "Expected I-beam joint topology"
+                assert topo_results.beam_a == beam_a, "Expected first beam as first plate in topology result"
+                assert topo_results.beam_b == beam_b, "Expected second beam as second plate in topology result"
+                assert TOL.is_close(topo_results.distance, 0.0), "Expected distance == 0"
+                assert isinstance(topo_results.location, Point), "Expected location to be a Point instance"
 
 
 def test_find_l_topology_tolerance_no_max_dist(l_beams_with_tolerance):
@@ -145,7 +145,7 @@ def test_find_l_topology_tolerance_no_max_dist(l_beams_with_tolerance):
             for beam_b in [bb, flip_beam(bb)]:
                 cs = ConnectionSolver()
                 topo_results = cs.find_topology(beam_a, beam_b)
-                assert topo_results[0] != JointTopology.TOPO_L, "Expected joint topology {}".format(JointTopology.get_name(topo_results[0]))
+                assert topo_results.topology != JointTopology.TOPO_L, "Expected joint topology {}".format(JointTopology.get_name(topo_results.topology))
 
 
 def test_find_topology_l_beams_with_max_dist(l_beams_with_tolerance):
@@ -155,11 +155,11 @@ def test_find_topology_l_beams_with_max_dist(l_beams_with_tolerance):
                 cs = ConnectionSolver()
                 dist = distance_segment_segment(beam_a.centerline, beam_b.centerline)
                 topo_results = cs.find_topology(beam_a, beam_b, max_distance=0.02)
-                assert topo_results[0] == JointTopology.TOPO_L, "Expected I-beam joint topology"
-                assert topo_results[1] == beam_a, "Expected first beam as first plate in topology result"
-                assert topo_results[2] == beam_b, "Expected second beam as second plate in topology result"
-                assert TOL.is_close(topo_results[3], dist), "Expected distance == 0.01"
-                assert isinstance(topo_results[4], Point), "Expected location to be a Point instance"
+                assert topo_results.topology == JointTopology.TOPO_L, "Expected I-beam joint topology"
+                assert topo_results.beam_a == beam_a, "Expected first beam as first plate in topology result"
+                assert topo_results.beam_b == beam_b, "Expected second beam as second plate in topology result"
+                assert TOL.is_close(topo_results.distance, dist), "Expected distance == 0.01"
+                assert isinstance(topo_results.location, Point), "Expected location to be a Point instance"
 
 
 @pytest.fixture
@@ -190,11 +190,11 @@ def test_find_topology_t_beams(t_beams):
             for beam_b in [bb, flip_beam(bb)]:
                 cs = ConnectionSolver()
                 topo_results = cs.find_topology(beam_a, beam_b)
-                assert topo_results[0] == JointTopology.TOPO_T, "Expected I-beam joint topology"
-                assert topo_results[1] == beam_a, "Expected first beam as first plate in topology result"
-                assert topo_results[2] == beam_b, "Expected second beam as second plate in topology result"
-                assert TOL.is_close(topo_results[3], 0.0), "Expected distance == 0"
-                assert isinstance(topo_results[4], Point), "Expected location to be a Point instance"
+                assert topo_results.topology == JointTopology.TOPO_T, "Expected I-beam joint topology"
+                assert topo_results.beam_a == beam_a, "Expected first beam as first plate in topology result"
+                assert topo_results.beam_b == beam_b, "Expected second beam as second plate in topology result"
+                assert TOL.is_close(topo_results.distance, 0.0), "Expected distance == 0"
+                assert isinstance(topo_results.location, Point), "Expected location to be a Point instance"
 
 
 def test_find_t_topology_tolerance_no_max_dist(t_beams_with_tolerance):
@@ -203,7 +203,7 @@ def test_find_t_topology_tolerance_no_max_dist(t_beams_with_tolerance):
             for beam_b in [bb, flip_beam(bb)]:
                 cs = ConnectionSolver()
                 topo_results = cs.find_topology(beam_a, beam_b)
-                assert topo_results[0] != JointTopology.TOPO_T, "Expected joint topology {}".format(JointTopology.get_name(topo_results[0]))
+                assert topo_results.topology != JointTopology.TOPO_T, "Expected joint topology {}".format(JointTopology.get_name(topo_results.topology))
 
 
 def test_find_topology_t_beams_with_max_dist(t_beams_with_tolerance):
@@ -213,11 +213,11 @@ def test_find_topology_t_beams_with_max_dist(t_beams_with_tolerance):
                 cs = ConnectionSolver()
                 dist = distance_segment_segment(beam_a.centerline, beam_b.centerline)
                 topo_results = cs.find_topology(beam_a, beam_b, max_distance=0.02)
-                assert topo_results[0] == JointTopology.TOPO_T, "Expected T-beam joint topology"
-                assert topo_results[1] == beam_a, "Expected first beam as first plate in topology result"
-                assert topo_results[2] == beam_b, "Expected second beam as second plate in topology result"
-                assert TOL.is_close(topo_results[3], dist), "Expected distance == 0.01"
-                assert isinstance(topo_results[4], Point), "Expected location to be a Point instance"
+                assert topo_results.topology == JointTopology.TOPO_T, "Expected T-beam joint topology"
+                assert topo_results.beam_a == beam_a, "Expected first beam as first plate in topology result"
+                assert topo_results.beam_b == beam_b, "Expected second beam as second plate in topology result"
+                assert TOL.is_close(topo_results.distance, dist), "Expected distance == 0.01"
+                assert isinstance(topo_results.location, Point), "Expected location to be a Point instance"
 
 
 def test_find_topology_t_beams_with_max_dist_flip_role(t_beams_with_tolerance):
@@ -227,11 +227,11 @@ def test_find_topology_t_beams_with_max_dist_flip_role(t_beams_with_tolerance):
                 cs = ConnectionSolver()
                 dist = distance_segment_segment(beam_a.centerline, beam_b.centerline)
                 topo_results = cs.find_topology(beam_a, beam_b, max_distance=0.02)
-                assert topo_results[0] == JointTopology.TOPO_T, "Expected T-beam joint topology"
-                assert topo_results[1] == beam_b, "Expected first beam as first plate in topology result"
-                assert topo_results[2] == beam_a, "Expected second beam as second plate in topology result"
-                assert TOL.is_close(topo_results[3], dist), "Expected distance == 0.01"
-                assert isinstance(topo_results[4], Point), "Expected location to be a Point instance"
+                assert topo_results.topology == JointTopology.TOPO_T, "Expected T-beam joint topology"
+                assert topo_results.beam_b == beam_a, "Expected first beam as first plate in topology result"
+                assert topo_results.beam_a == beam_b, "Expected second beam as second plate in topology result"
+                assert TOL.is_close(topo_results.distance, dist), "Expected distance == 0.01"
+                assert isinstance(topo_results.location, Point), "Expected location to be a Point instance"
 
 
 @pytest.fixture
@@ -264,11 +264,11 @@ def test_plate_L_topos():
     cs = PlateConnectionSolver()
 
     topo_results = cs.find_topology(plate_a, plate_b)
-    assert topo_results[0] == JointTopology.TOPO_EDGE_EDGE, "Expected L-joint topology"
-    assert topo_results[1][0] == plate_a, "Expected plate_a as first plate in topology result"
-    assert topo_results[2][0] == plate_b, "Expected plate_b as second plate in topology result"
-    assert topo_results[1][1] == 1, "Expected connection segment at index = 1"
-    assert topo_results[2][1] == 0, "Expected connection segment at index = 0"
+    assert topo_results.topology == JointTopology.TOPO_EDGE_EDGE, "Expected L-joint topology"
+    assert topo_results.plate_a == plate_a, "Expected plate_a as first plate in topology result"
+    assert topo_results.plate_b == plate_b, "Expected plate_b as second plate in topology result"
+    assert topo_results.segment_a_index == 1, "Expected connection segment at index = 1"
+    assert topo_results.segment_b_index == 0, "Expected connection segment at index = 0"
 
 
 def test_plate_T_topos():
@@ -283,11 +283,11 @@ def test_plate_T_topos():
     cs = PlateConnectionSolver()
 
     topo_results = cs.find_topology(plate_a, plate_b)
-    assert topo_results[0] == JointTopology.TOPO_EDGE_FACE, "Expected T-joint topology"
-    assert topo_results[1][0] == plate_b, "Expected plate_a as first plate in topology result"
-    assert topo_results[2][0] == plate_a, "Expected plate_b as second plate in topology result"
-    assert topo_results[1][1] == 0, "Expected connection segment at index = 1"
-    assert topo_results[2][1] is None, "Expected connection segment at index = 0"
+    assert topo_results.topology == JointTopology.TOPO_EDGE_FACE, "Expected T-joint topology"
+    assert topo_results.plate_a == plate_b, "Expected plate_a as first plate in topology result"
+    assert topo_results.plate_b == plate_a, "Expected plate_b as second plate in topology result"
+    assert topo_results.segment_a_index == 0, "Expected connection segment at index = 1"
+    assert topo_results.segment_b_index is None, "Expected connection segment at index = 0"
 
 
 def test_reversed_plate_T_topos():
@@ -302,11 +302,11 @@ def test_reversed_plate_T_topos():
     cs = PlateConnectionSolver()
 
     topo_results = cs.find_topology(plate_b, plate_a)
-    assert topo_results[0] == JointTopology.TOPO_EDGE_FACE, "Expected T-joint topology"
-    assert topo_results[1][0] == plate_b, "Expected plate_a as first plate in topology result"
-    assert topo_results[2][0] == plate_a, "Expected plate_b as second plate in topology result"
-    assert topo_results[1][1] == 0, "Expected connection segment at index = 1"
-    assert topo_results[2][1] is None, "Expected connection segment at index = 0"
+    assert topo_results.topology == JointTopology.TOPO_EDGE_FACE, "Expected T-joint topology"
+    assert topo_results.plate_a == plate_b, "Expected plate_a as first plate in topology result"
+    assert topo_results.plate_b == plate_a, "Expected plate_b as second plate in topology result"
+    assert topo_results.segment_a_index == 0, "Expected connection segment at index = 1"
+    assert topo_results.segment_b_index is None, "Expected connection segment at index = 0"
 
 
 def test_three_plate_topos():
@@ -331,7 +331,7 @@ def test_three_plate_topos():
     topo_results.append(cs.find_topology(plate_a, plate_c))
 
     assert len(topo_results) == 3, "Expected three topology results"
-    assert all(tr[0] == JointTopology.TOPO_EDGE_EDGE for tr in topo_results), "Expected all topology results to be L-joints"
+    assert all(tr.topology == JointTopology.TOPO_EDGE_EDGE for tr in topo_results), "Expected all topology results to be L-joints"
 
 
 def test_three_plate_mix_topos():
@@ -356,6 +356,6 @@ def test_three_plate_mix_topos():
     topo_results.append(cs.find_topology(plate_a, plate_c))
 
     assert len(topo_results) == 3, "Expected three topology results"
-    assert topo_results[0][0] == JointTopology.TOPO_EDGE_FACE, "Expected first topology result to be T-joint"
-    assert topo_results[1][0] == JointTopology.TOPO_EDGE_EDGE, "Expected second topology result to be L-joint"
-    assert topo_results[2][0] == JointTopology.TOPO_EDGE_EDGE, "Expected third topology result to be L-joint"
+    assert topo_results[0].topology == JointTopology.TOPO_EDGE_FACE, "Expected first topology result to be T-joint"
+    assert topo_results[1].topology == JointTopology.TOPO_EDGE_EDGE, "Expected second topology result to be L-joint"
+    assert topo_results[2].topology == JointTopology.TOPO_EDGE_EDGE, "Expected third topology result to be L-joint"